+++ interfaces/interface.py
@@ -72,6 +72,16 @@ def configure_interfaces(config):
                 logger.error(f"Unable to configure interface {i}: {repr(e)}")
                 raise
 
+        elif interface_type == "loraham":
+            try:
+                from . import lorahaminterface
+                i_face = lorahaminterface.LoRaHAMInterface(data)
+
+                interfaces.append(i_face)
+            except Exception as e:
+                logger.error(f"Unable to configure interface {i}: {repr(e)}")
+                raise
+
         else:
             logger.error(f"Interface {i} is unknown type {interface_type}")